refactor: move the series repository to the module
Move SeriesRepository into the SeriesModule, will try to ensure all series interaction is done throu
gh this
