+++ app/src/main/java/com/chesire/malime/flow/profile/ProfileViewModel.kt
@@ -4,12 +4,11 @@ import androidx.lifecycle.Transformations
 import androidx.lifecycle.ViewModel
 import com.chesire.malime.core.flags.UserSeriesStatus
 import com.chesire.malime.core.models.SeriesModel
-import com.chesire.malime.repo.SeriesRepository
-import com.chesire.malime.account.UserRepository
+import com.chesire.malime.series.SeriesRepository
 import javax.inject.Inject
 
 class ProfileViewModel @Inject constructor(
-    seriesRepository: SeriesRepository,
+    seriesRepository: com.chesire.malime.series.SeriesRepository,
     userRepository: com.chesire.malime.account.UserRepository
 ) : ViewModel() {
     val user = userRepository.user
