+++ app/src/main/java/com/chesire/malime/flow/series/list/SeriesListViewModel.kt
@@ -6,7 +6,7 @@ import com.chesire.malime.AuthCaster
 import com.chesire.malime.server.Resource
 import com.chesire.malime.core.flags.UserSeriesStatus
 import com.chesire.malime.core.models.SeriesModel
-import com.chesire.malime.repo.SeriesRepository
+import com.chesire.malime.series.SeriesRepository
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
@@ -14,7 +14,7 @@ import javax.inject.Inject
  * ViewModel to use with the [SeriesListFragment], handles sending updates for a series.
  */
 class SeriesListViewModel @Inject constructor(
-    private val repo: SeriesRepository,
+    private val repo: com.chesire.malime.series.SeriesRepository,
     private val authCaster: AuthCaster
 ) : ViewModel() {
     val animeSeries = repo.anime
