+++ app/src/main/java/com/chesire/malime/services/RefreshSeriesWorker.kt
@@ -5,8 +5,7 @@ import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
 import com.chesire.malime.MalimeApplication
 import com.chesire.malime.server.Resource
-import com.chesire.malime.repo.SeriesRepository
-import com.chesire.malime.account.UserRepository
+import com.chesire.malime.series.SeriesRepository
 import timber.log.Timber
 import javax.inject.Inject
 
@@ -16,7 +15,7 @@ class RefreshSeriesWorker(
 ) : CoroutineWorker(appContext, workerParams) {
 
     @Inject
-    lateinit var seriesRepo: SeriesRepository
+    lateinit var seriesRepo: com.chesire.malime.series.SeriesRepository
     @Inject
     lateinit var userRepo: com.chesire.malime.account.UserRepository
