+++ app/src/test/java/com/chesire/malime/flow/profile/ProfileViewModelTests.kt
@@ -1,12 +1,8 @@
 package com.chesire.malime.flow.profile
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
-import com.chesire.malime.core.flags.UserSeriesStatus
-import com.chesire.malime.createSeriesModel
-import com.chesire.malime.repo.SeriesRepository
-import com.chesire.malime.account.UserRepository
+import com.chesire.malime.series.SeriesRepository
 import io.mockk.Runs
 import io.mockk.every
 import io.mockk.just
@@ -22,7 +18,7 @@ class ProfileViewModelTests {
 
     @Test
     fun `anime gets updated with seriesRepository anime series`() {
-        val mockSeriesRepo = mockk<SeriesRepository> {
+        val mockSeriesRepo = mockk<com.chesire.malime.series.SeriesRepository> {
             every { anime } returns MutableLiveData(listOf(createSeriesModel()))
             every { manga } returns mockk()
         }
@@ -41,7 +37,7 @@ class ProfileViewModelTests {
 
     @Test
     fun `manga gets updated with seriesRepository manga series`() {
-        val mockSeriesRepo = mockk<SeriesRepository> {
+        val mockSeriesRepo = mockk<com.chesire.malime.series.SeriesRepository> {
             every { anime } returns mockk()
             every { manga } returns MutableLiveData(listOf(createSeriesModel()))
         }
@@ -60,7 +56,7 @@ class ProfileViewModelTests {
 
     @Test
     fun `series has expected total items`() {
-        val mockSeriesRepo = mockk<SeriesRepository> {
+        val mockSeriesRepo = mockk<com.chesire.malime.series.SeriesRepository> {
             every { anime } returns mockk()
             every { manga } returns MutableLiveData(
                 listOf(
@@ -86,7 +82,7 @@ class ProfileViewModelTests {
 
     @Test
     fun `series has expected current items`() {
-        val mockSeriesRepo = mockk<SeriesRepository> {
+        val mockSeriesRepo = mockk<com.chesire.malime.series.SeriesRepository> {
             every { anime } returns mockk()
             every { manga } returns MutableLiveData(
                 listOf(
@@ -112,7 +108,7 @@ class ProfileViewModelTests {
 
     @Test
     fun `series has expected completed items`() {
-        val mockSeriesRepo = mockk<SeriesRepository> {
+        val mockSeriesRepo = mockk<com.chesire.malime.series.SeriesRepository> {
             every { anime } returns mockk()
             every { manga } returns MutableLiveData(
                 listOf(
@@ -138,7 +134,7 @@ class ProfileViewModelTests {
 
     @Test
     fun `series has expected onHold items`() {
-        val mockSeriesRepo = mockk<SeriesRepository> {
+        val mockSeriesRepo = mockk<com.chesire.malime.series.SeriesRepository> {
             every { anime } returns mockk()
             every { manga } returns MutableLiveData(
                 listOf(
@@ -164,7 +160,7 @@ class ProfileViewModelTests {
 
     @Test
     fun `series has expected dropped items`() {
-        val mockSeriesRepo = mockk<SeriesRepository> {
+        val mockSeriesRepo = mockk<com.chesire.malime.series.SeriesRepository> {
             every { anime } returns mockk()
             every { manga } returns MutableLiveData(
                 listOf(
@@ -190,7 +186,7 @@ class ProfileViewModelTests {
 
     @Test
     fun `series has expected planned items`() {
-        val mockSeriesRepo = mockk<SeriesRepository> {
+        val mockSeriesRepo = mockk<com.chesire.malime.series.SeriesRepository> {
             every { anime } returns mockk()
             every { manga } returns MutableLiveData(
                 listOf(
@@ -216,7 +212,7 @@ class ProfileViewModelTests {
 
     @Test
     fun `series has expected unknown items`() {
-        val mockSeriesRepo = mockk<SeriesRepository> {
+        val mockSeriesRepo = mockk<com.chesire.malime.series.SeriesRepository> {
             every { anime } returns mockk()
             every { manga } returns MutableLiveData(
                 listOf(
