+++ app/src/test/java/com/chesire/malime/flow/series/list/SeriesListViewModelTests.kt
@@ -3,11 +3,9 @@ package com.chesire.malime.flow.series.list
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import com.chesire.malime.AuthCaster
 import com.chesire.malime.CoroutinesMainDispatcherRule
-import com.chesire.malime.server.Resource
 import com.chesire.malime.core.flags.UserSeriesStatus
-import com.chesire.malime.repo.SeriesRepository
+import com.chesire.malime.series.SeriesRepository
 import io.mockk.Runs
-import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.every
 import io.mockk.just
@@ -25,7 +23,7 @@ class SeriesListViewModelTests {
 
     @Test
     fun `updateSeries sends request through the repository`() {
-        val mockRepo = mockk<SeriesRepository> {
+        val mockRepo = mockk<com.chesire.malime.series.SeriesRepository> {
             coEvery {
                 updateSeries(0, 0, UserSeriesStatus.Current)
             } coAnswers {
@@ -44,7 +42,7 @@ class SeriesListViewModelTests {
 
     @Test
     fun `updateSeries 401 failure notifies through AuthCaster`() {
-        val mockRepo = mockk<SeriesRepository> {
+        val mockRepo = mockk<com.chesire.malime.series.SeriesRepository> {
             coEvery {
                 updateSeries(0, 0, UserSeriesStatus.Current)
             } coAnswers {
@@ -66,7 +64,7 @@ class SeriesListViewModelTests {
     @Test
     fun `updateSeries failure not 401 invokes callback`() {
         var condition = false
-        val mockRepo = mockk<SeriesRepository> {
+        val mockRepo = mockk<com.chesire.malime.series.SeriesRepository> {
             coEvery {
                 updateSeries(0, 0, UserSeriesStatus.Current)
             } coAnswers {
@@ -86,7 +84,7 @@ class SeriesListViewModelTests {
     @Test
     fun `updateSeries success invokes callback`() {
         var condition = false
-        val mockRepo = mockk<SeriesRepository> {
+        val mockRepo = mockk<com.chesire.malime.series.SeriesRepository> {
             coEvery {
                 updateSeries(0, 0, UserSeriesStatus.Current)
             } coAnswers {
