+++ app/src/test/java/com/chesire/malime/flow/series/search/SearchViewModelTests.kt
@@ -2,14 +2,13 @@ package com.chesire.malime.flow.series.search
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
-import com.chesire.malime.core.flags.AsyncState
 import com.chesire.malime.AuthCaster
 import com.chesire.malime.CoroutinesMainDispatcherRule
 import com.chesire.malime.server.Resource
 import com.chesire.malime.server.api.SearchApi
 import com.chesire.malime.core.flags.SeriesType
 import com.chesire.malime.core.models.SeriesModel
-import com.chesire.malime.repo.SeriesRepository
+import com.chesire.malime.series.SeriesRepository
 import io.mockk.Runs
 import io.mockk.coEvery
 import io.mockk.coVerify
@@ -28,7 +27,7 @@ class SearchViewModelTests {
 
     @Test
     fun `performSearch with no search title posts error`() {
-        val mockRepo = mockk<SeriesRepository>()
+        val mockRepo = mockk<com.chesire.malime.series.SeriesRepository>()
         val mockSearch = mockk<SearchApi>()
         val mockObserver = mockk<Observer<com.chesire.malime.core.flags.AsyncState<List<SeriesModel>, SearchError>>> {
             every { onChanged(any()) } just Runs
@@ -45,7 +44,7 @@ class SearchViewModelTests {
 
     @Test
     fun `performSearch with seriesType of anime uses correct api`() {
-        val mockRepo = mockk<SeriesRepository>()
+        val mockRepo = mockk<com.chesire.malime.series.SeriesRepository>()
         val mockSearch = mockk<SearchApi> {
             coEvery { searchForAnime(any()) } returns Resource.Error("", 0)
         }
@@ -65,7 +64,7 @@ class SearchViewModelTests {
 
     @Test
     fun `performSearch with seriesType of manga uses correct api`() {
-        val mockRepo = mockk<SeriesRepository>()
+        val mockRepo = mockk<com.chesire.malime.series.SeriesRepository>()
         val mockSearch = mockk<SearchApi> {
             coEvery { searchForManga(any()) } returns Resource.Error("", 0)
         }
@@ -87,7 +86,7 @@ class SearchViewModelTests {
     @Test
     fun `performSearch on success posts success with data`() {
         val expected = listOf<SeriesModel>(mockk())
-        val mockRepo = mockk<SeriesRepository>()
+        val mockRepo = mockk<com.chesire.malime.series.SeriesRepository>()
         val mockSearch = mockk<SearchApi> {
             coEvery {
                 searchForAnime(any())
