+++ series/build.gradle
@@ -34,6 +34,7 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
+    implementation project(path: ':account')
     implementation project(path: ':core')
     implementation project(path: ':database')
     implementation project(path: ':server')
