+++ series/src/main/AndroidManifest.xml
@@ -1,2 +1,1 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.chesire.malime.series" />
+<manifest package="com.chesire.malime.series" />
